ci: speed up regular test workflow (#8812)
* ci: speed up regular test workflow

Reuse expensive test runtimes, parallelize Javadoc reactor work, and run trusted builds on the 8-core runner.


* test(http): keep process assertion on test thread


---------

(cherry picked from commit baddb284453516a12e7e032dada8647877616241)
