+++ .github/workflows/TEST_FEATURE_BRANCH.yml
@@ -22,7 +22,7 @@ concurrency:
 jobs:
   run-tests:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ github.repository_owner == 'camunda' && (github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false) && 'gcp-core-8-default' || 'ubuntu-latest' }}
     steps:
       - uses: actions/checkout@v6
         with:
@@ -111,7 +111,7 @@ jobs:
 
       # Build and test (format and javadoc run in parallel jobs)
       - name: Build Connectors
-        run: ./mvnw --batch-mode -U clean test -Dquickly -T 1C
+        run: ./mvnw --batch-mode -U clean test -Dquickly -T 1C '-P!autoFormat'
 
       - name: Publish Test Report
         if: always() && steps.internal.outputs.internal == 'true'
@@ -153,7 +153,7 @@ jobs:
 
   check-javadoc:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ github.repository_owner == 'camunda' && (github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false) && 'gcp-core-8-default' || 'ubuntu-latest' }}
     steps:
       - uses: actions/checkout@v6
         with:
@@ -227,14 +227,14 @@ jobs:
           mirrors: '[{"url": "https://repository.nexus.camunda.cloud/content/groups/internal/", "id": "camunda-nexus", "mirrorOf": "*,!confluent,!shibboleth", "name": "camunda Nexus"}]'
 
       - name: Install element-template-generator plugin
-        run: ./mvnw --batch-mode install -pl element-template-generator/maven-plugin -am -DskipTests -Dquickly
+        run: ./mvnw --batch-mode install -pl element-template-generator/maven-plugin -am -DskipTests -Dquickly -T 1C
 
       - name: Check Javadoc
-        run: ./mvnw --batch-mode compile javadoc:javadoc -Dquickly
+        run: ./mvnw --batch-mode compile javadoc:javadoc -Dquickly -T 1C
 
   check-format:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ github.repository_owner == 'camunda' && (github.event_name != 'pull_request' || github.event.pull_request.head.repo.fork == false) && 'gcp-core-8-default' || 'ubuntu-latest' }}
     steps:
       - uses: actions/checkout@v6
         with:
@@ -308,7 +308,7 @@ jobs:
           mirrors: '[{"url": "https://repository.nexus.camunda.cloud/content/groups/internal/", "id": "camunda-nexus", "mirrorOf": "*,!confluent,!shibboleth", "name": "camunda Nexus"}]'
 
       - name: Install element-template-generator plugin
-        run: ./mvnw --batch-mode install -pl element-template-generator/maven-plugin -am -DskipTests -Dquickly
+        run: ./mvnw --batch-mode install -pl element-template-generator/maven-plugin -am -DskipTests -Dquickly -T 1C
 
       - name: Check Format
         run: ./mvnw --batch-mode validate -PcheckFormat -Dquickly
